+++ doc/CHANGES.rst
@@ -153,6 +153,8 @@ Features Added
 Bugs Fixed
 ++++++++++
 
+- LP #143655:  don't prevent sorting using a path index.
+
 - LP #142478:  normalize terms passed to ``PLexicon.queryLexicon`` using
   the lexicon's pipeline (e.g., case flattening, stop word removal, etc.)